config: don't crash if config file doesn't exist.

+++ lib/config.py
@@ -75,7 +75,7 @@ class Config(object):
             os.fsync(file.fileno())
         finally:
             file.close()
-        if os.name == 'nt':
+        if os.name == 'nt' and os.path.exists(path):
             # Windows doesn't support atomic renames.
             backup_path = path + '.bak'
             os.rename(path, backup_path)
